Moved lagrange_interp and finite_differencing into utils
