+++ src/utils/finite_differencing.cpp
