+++ src/utils/lagrange_interp.hpp
@@ -14,7 +14,7 @@
 #include <cstring>
 #include <new>
 
-#include "athena.hpp"
+#include "../athena.hpp"
 
 // If this is uncommented always use symmetric operators
 #ifndef LOCALINTERP_SYMMETRIC
