+++ src/wave/wave.hpp
@@ -12,7 +12,7 @@
 #include "../athena.hpp"
 #include "../athena_arrays.hpp"
 
-#include "../finite_differencing.hpp"
+#include "../utils/finite_differencing.hpp"
 #ifdef COMPACT_FD
 #include "../finite_difference/compact_finite_differencing.hpp"
 #endif // COMPACT_FD
