+++ src/z4c/z4c.hpp
@@ -20,7 +20,7 @@
 #include "../athena.hpp"
 #include "../athena_arrays.hpp"
 #include "../athena_tensor.hpp"
-#include "../finite_differencing.hpp"
+#include "../utils/finite_differencing.hpp"
 
 #include "../bvals/cc/bvals_cc.hpp"
 #include "../bvals/vc/bvals_vc.hpp"
